# services/marketing-ops/openapi/marketing-ops.v1.yaml
openapi: 3.1.0
info:
  title: Nexus Marketing Ops API
  version: 1.0.0
  description: REST contract for the operational campaign workspace.
servers:
  - url: /v1
security:
  - bearerAuth: []
paths:
  /ads-integrations/{provider}/setup:
    parameters:
      - name: provider
        in: path
        required: true
        schema: { $ref: '#/components/schemas/AdsProvider' }
    get:
      operationId: readAdsSetup
      summary: Read safe installation application metadata
      description: Administrator only, verified against the canonical application identity. Never returns credentials or tokens.
      responses:
        '200':
          description: Safe setup state; prepared does not mean authorized.
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { type: object, additionalProperties: false, required: [data], properties: { data: { $ref: '#/components/schemas/AdsSetup' } } }
        default: { $ref: '#/components/responses/Error' }
    post:
      operationId: saveAdsSetup
      summary: Configure or replace the installation application
      description: Administrator only. Secrets are stored privately on the server. Replacement requires explicit confirmation and invalidates prior authorizations while preserving history. Empty secret preserves an existing managed secret; initial setup and external takeover require a new secret. Callback origin comes exclusively from trusted server configuration.
      parameters:
        - { $ref: '#/components/parameters/IfMatch' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdsSetupInput' }
      responses:
        '200':
          description: Safe metadata for the published configuration.
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { type: object, additionalProperties: false, required: [data], properties: { data: { $ref: '#/components/schemas/AdsSetup' } } }
        default: { $ref: '#/components/responses/Error' }
  /ads-integrations: {"get":{"operationId":"listAdsIntegrations","summary":"Read safe connection states","parameters":[],"responses":{"200":{"description":"Read safe connection states","content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"required":["data"],"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/AdsConnection"}}}}}}},"default":{"$ref":"#/components/responses/Error"}}}}
  /ads-integrations/{provider}/authorize: {"post":{"operationId":"authorizeAds","summary":"Manage the installation's Ads authorization","parameters":[{"name":"provider","in":"path","required":true,"schema":{"$ref":"#/components/schemas/AdsProvider"}},{"name":"X-ENS-OAuth-Session","in":"header","required":true,"description":"BFF-derived HMAC of authenticated session, never a browser-supplied identity.","schema":{"type":"string","pattern":"^[a-f0-9]{64}$"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdsEmpty"}}}},"responses":{"200":{"description":"Manage the installation's Ads authorization","content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"required":["data"],"properties":{"data":{"$ref":"#/components/schemas/AdsAuthorization"}}}}}},"default":{"$ref":"#/components/responses/Error"}}}}
  /ads-integrations/{provider}/callback: {"post":{"operationId":"completeAdsAuthorization","summary":"Manage the installation's Ads authorization","parameters":[{"name":"provider","in":"path","required":true,"schema":{"$ref":"#/components/schemas/AdsProvider"}},{"name":"X-ENS-OAuth-Session","in":"header","required":true,"description":"BFF-derived HMAC of authenticated session, never a browser-supplied identity.","schema":{"type":"string","pattern":"^[a-f0-9]{64}$"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdsCallback"}}}},"responses":{"200":{"description":"Manage the installation's Ads authorization","content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"required":["data"],"properties":{"data":{"$ref":"#/components/schemas/AdsConnection"}}}}}},"default":{"$ref":"#/components/responses/Error"}}}}
  /ads-integrations/{provider}/account: {"post":{"operationId":"selectAdsAccount","summary":"Manage the installation's Ads authorization","parameters":[{"name":"provider","in":"path","required":true,"schema":{"$ref":"#/components/schemas/AdsProvider"}},{"$ref":"#/components/parameters/IfMatch"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdsAccountInput"}}}},"responses":{"200":{"description":"Manage the installation's Ads authorization","headers":{"ETag":{"$ref":"#/components/headers/ETag"}},"content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"required":["data"],"properties":{"data":{"$ref":"#/components/schemas/AdsConnection"}}}}}},"default":{"$ref":"#/components/responses/Error"}}}}
  /ads-integrations/{provider}/disconnect: {"post":{"operationId":"disconnectAds","summary":"Manage the installation's Ads authorization","parameters":[{"name":"provider","in":"path","required":true,"schema":{"$ref":"#/components/schemas/AdsProvider"}},{"$ref":"#/components/parameters/IfMatch"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdsEmpty"}}}},"responses":{"200":{"description":"Manage the installation's Ads authorization","headers":{"ETag":{"$ref":"#/components/headers/ETag"}},"content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"required":["data"],"properties":{"data":{"$ref":"#/components/schemas/AdsConnection"}}}}}},"default":{"$ref":"#/components/responses/Error"}}}}
  /ads-integrations/{provider}/accounts: {"get":{"operationId":"listAdsAccounts","summary":"Read accessible advertising accounts","parameters":[{"name":"provider","in":"path","required":true,"schema":{"$ref":"#/components/schemas/AdsProvider"}}],"responses":{"200":{"description":"Read accessible advertising accounts","content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"required":["data"],"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/AdsAccount"}}}}}}},"default":{"$ref":"#/components/responses/Error"}}}}
  /ads-integrations/{provider}/campaigns: {"get":{"operationId":"listExternalAdsCampaigns","summary":"Read campaigns from the selected advertising account","parameters":[{"name":"provider","in":"path","required":true,"schema":{"$ref":"#/components/schemas/AdsProvider"}}],"responses":{"200":{"description":"Read campaigns from the selected advertising account","content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"required":["data"],"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/AdsExternalCampaign"}}}}}}},"default":{"$ref":"#/components/responses/Error"}}}}
  /campaigns/{id}/ads-links: {"get":{"operationId":"listAdsLinks","summary":"Read authorized campaign links","parameters":[{"$ref":"#/components/parameters/Id"}],"responses":{"200":{"description":"Read authorized campaign links","content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"required":["data"],"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/AdsLink"}}}}}}},"default":{"$ref":"#/components/responses/Error"}}},"post":{"operationId":"createAdsLink","summary":"Confirm advertisement attribution","parameters":[{"$ref":"#/components/parameters/Id"},{"$ref":"#/components/parameters/IdempotencyKey"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdsLinkInput"}}}},"responses":{"201":{"description":"Confirm advertisement attribution","headers":{"ETag":{"$ref":"#/components/headers/ETag"}},"content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"required":["data"],"properties":{"data":{"$ref":"#/components/schemas/AdsLink"}}}}}},"default":{"$ref":"#/components/responses/Error"}}}}
  /campaigns/{id}/ads-links/{linkId}/disable: {"post":{"operationId":"disableAdsLink","summary":"Stop further synchronization and preserve history","parameters":[{"$ref":"#/components/parameters/Id"},{"name":"linkId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"$ref":"#/components/parameters/IfMatch"},{"$ref":"#/components/parameters/IdempotencyKey"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdsEmpty"}}}},"responses":{"200":{"description":"Stop further synchronization and preserve history","headers":{"ETag":{"$ref":"#/components/headers/ETag"}},"content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"required":["data"],"properties":{"data":{"$ref":"#/components/schemas/AdsLink"}}}}}},"default":{"$ref":"#/components/responses/Error"}}}}
  /campaigns/{id}/ads-links/{linkId}/sync: {"post":{"operationId":"syncAdsLink","summary":"Read provider data with reviewed native contact batches","parameters":[{"$ref":"#/components/parameters/Id"},{"name":"linkId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"$ref":"#/components/parameters/IdempotencyKey"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdsPeriod"}}}},"responses":{"200":{"description":"Read provider data with reviewed native contact batches","content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"required":["data"],"properties":{"data":{"$ref":"#/components/schemas/AdsSyncReceipt"}}}}}},"default":{"$ref":"#/components/responses/Error"}}}}
  /campaigns/{id}/ads-links/{linkId}/results: {"get":{"operationId":"readAdsResults","summary":"Read provider measurements and durable receipts","parameters":[{"$ref":"#/components/parameters/Id"},{"name":"linkId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Read provider measurements and durable receipts","content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"required":["data"],"properties":{"data":{"$ref":"#/components/schemas/AdsResults"}}}}}},"default":{"$ref":"#/components/responses/Error"}}}}
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  headers:
    ETag:
      description: Quoted current resource or aggregate version.
      schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 1, maxLength: 128 } }
    IfMatch: { name: If-Match, in: header, required: true, schema: { type: string, pattern: '^(W/)?("[1-9][0-9]*"|[1-9][0-9]*)$' } }
  responses:
    Error:
      description: Stable public error envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    AdsSetup:
      type: object
      additionalProperties: false
      required: [provider, version, mode, writable, ready, publicOrigin, redirectUri, clientId, apiVersion, scopes, metaLoginConfigId, googleLoginCustomerId, hasClientSecret]
      properties:
        provider: { $ref: '#/components/schemas/AdsProvider' }
        version: { type: integer, minimum: 1 }
        mode: { enum: [empty, managed, external] }
        writable: { type: boolean }
        ready: { type: boolean, description: Includes the current trusted installation origin; persisted credentials alone do not imply readiness. }
        publicOrigin: { type: [string, 'null'], format: uri }
        redirectUri: { type: [string, 'null'], format: uri }
        clientId: { type: [string, 'null'] }
        apiVersion: { type: [string, 'null'] }
        scopes: { type: array, items: { type: string } }
        metaLoginConfigId: { type: [string, 'null'] }
        googleLoginCustomerId: { type: [string, 'null'] }
        hasClientSecret: { type: boolean }
    AdsSetupInput:
      type: object
      additionalProperties: false
      required: [clientId, apiVersion, scopes]
      description: Provider-specific validation rejects unrelated fields. Meta requires metaLoginConfigId and vXX.0; Google accepts googleLoginCustomerId and vXX; LinkedIn uses YYYYMM. Permissions must belong to the provider's allowlist. Google API access belongs to the Cloud project that owns the OAuth credentials.
      properties:
        clientId: { type: string, minLength: 1, maxLength: 256, pattern: '^[A-Za-z0-9._:-]+$' }
        clientSecret: { type: string, maxLength: 4096, writeOnly: true }
        apiVersion: { type: string, minLength: 1, maxLength: 12 }
        scopes: { type: array, minItems: 1, maxItems: 10, uniqueItems: true, items: { type: string, maxLength: 100 } }
        metaLoginConfigId: { type: string, pattern: '^\d{1,30}$' }
        googleLoginCustomerId: { type: string, description: Optional ten-digit manager ID; dashes are accepted and removed by the server. }
        confirmReplacement: { type: boolean }
        takeOverExternal: { type: boolean }
    AdsProvider: {"enum":["meta","google","linkedin"]}
    AdsEmpty: {"type":"object","additionalProperties":false,"required":[],"properties":{}}
    AdsAccount: {"type":"object","additionalProperties":false,"required":["id","name","currency","timeZone"],"properties":{"id":{"type":"string"},"name":{"type":"string"},"currency":{"type":"string","pattern":"^[A-Z]{3}$"},"timeZone":{"type":["string","null"]},"manager":{"type":"boolean"},"loginCustomerId":{"type":"string"}}}
    AdsConnection: {"type":"object","additionalProperties":false,"required":["provider","status","version","accounts","selectedAccountId","capabilities","lastSyncAt","safeError"],"properties":{"provider":{"$ref":"#/components/schemas/AdsProvider"},"status":{"enum":["unprepared","prepared","pending_account","connected","partial","reconnect_required","disconnected","error"]},"version":{"type":"integer","minimum":1},"accounts":{"type":"array","items":{"$ref":"#/components/schemas/AdsAccount"}},"selectedAccountId":{"type":["string","null"]},"capabilities":{"type":"object","additionalProperties":false,"required":["metrics","nativeLeads"],"properties":{"metrics":{"type":"boolean"},"nativeLeads":{"type":"boolean"}}},"lastSyncAt":{"type":["string","null"],"format":"date-time"},"safeError":{"type":["string","null"]}}}
    AdsAuthorization: {"type":"object","additionalProperties":false,"required":["authorizationUrl"],"properties":{"authorizationUrl":{"type":"string","format":"uri","description":"Fixed provider consent URL; no application secret."}}}
    AdsCallback: {"type":"object","additionalProperties":false,"required":["state"],"properties":{"state":{"type":"string","pattern":"^[A-Za-z0-9_-]{43}$"},"code":{"type":"string","minLength":1,"maxLength":4096},"error":{"type":"string","minLength":1,"maxLength":200}},"oneOf":[{"required":["code"]},{"required":["error"]}]}
    AdsAccountInput: {"type":"object","additionalProperties":false,"required":["accountId"],"properties":{"accountId":{"type":"string","minLength":1,"maxLength":200}}}
    AdsExternalCampaign: {"type":"object","additionalProperties":false,"required":["id","name","status"],"properties":{"id":{"type":"string"},"name":{"type":"string"},"status":{"type":"string"}}}
    AdsLinkInput: {"type":"object","additionalProperties":false,"required":["provider","sourceId","externalCampaignId","destination"],"properties":{"provider":{"$ref":"#/components/schemas/AdsProvider"},"sourceId":{"type":"string","format":"uuid"},"externalCampaignId":{"type":"string","minLength":1,"maxLength":200},"destination":{"enum":["native_form","landing_page","whatsapp"]}}}
    AdsLink: {"type":"object","additionalProperties":false,"required":["id","campaignId","provider","sourceId","externalAccountId","externalCampaignId","destination","enabled","version","lastSyncAt","safeError"],"properties":{"id":{"type":"string","format":"uuid"},"campaignId":{"type":"string","format":"uuid"},"provider":{"$ref":"#/components/schemas/AdsProvider"},"sourceId":{"type":"string","format":"uuid"},"externalAccountId":{"type":"string"},"externalCampaignId":{"type":"string"},"destination":{"enum":["native_form","landing_page","whatsapp"]},"enabled":{"type":"boolean"},"version":{"type":"integer","minimum":1},"lastSyncAt":{"type":["string","null"],"format":"date-time"},"safeError":{"type":["string","null"]}}}
    AdsPeriod: {"type":"object","additionalProperties":false,"required":["from","to"],"properties":{"from":{"type":"string","format":"date"},"to":{"type":"string","format":"date"}},"description":"Ordered, at most 30 days, no future day; provider account calendar."}
    AdsDailyMetric: {"type":"object","additionalProperties":false,"required":["date","currency","timeZone","spend","impressions","clicks","conversions"],"properties":{"date":{"type":"string","format":"date"},"currency":{"type":"string","pattern":"^[A-Z]{3}$"},"timeZone":{"type":"string"},"spend":{"type":"number","minimum":0},"impressions":{"type":"integer","minimum":0},"clicks":{"type":"integer","minimum":0},"conversions":{"type":["number","null"],"minimum":0,"description":"Provider conversion events, never confirmed sales."}}}
    AdsSyncReceipt: {"type":"object","additionalProperties":false,"required":["id","status","days","reports","previewId","previewIds","warnings","completedAt"],"properties":{"id":{"type":"string","format":"uuid"},"status":{"enum":["completed","needs_review"]},"days":{"type":"integer","minimum":0},"reports":{"type":"integer","minimum":0},"previewId":{"type":["string","null"],"format":"uuid"},"previewIds":{"type":"array","items":{"type":"string","format":"uuid"}},"warnings":{"type":"array","items":{"type":"string"}},"completedAt":{"type":"string","format":"date-time"}}}
    AdsResults: {"type":"object","additionalProperties":false,"required":["daily","receipts"],"properties":{"daily":{"type":"array","items":{"$ref":"#/components/schemas/AdsDailyMetric"}},"receipts":{"type":"array","items":{"$ref":"#/components/schemas/AdsSyncReceipt"}}}}
...
    PublicErrorCode:
      type: string
      enum:
        - validation_error
        - unauthorized
        - auth_unavailable
        - tenant_required
        - forbidden
        - not_found
        - origin_forbidden
        - feature_disabled
        - idempotency_key_required
        - precondition_required
        - payload_too_large
        - idempotency_conflict
        - idempotency_in_progress
        - approval_conflict
        - approval_expired
        - approval_target_changed
        - self_approval_forbidden
        - invalid_supersession
        - version_conflict
        - invalid_transition
        - campaign_requirements_missing
        - campaign_read_only
        - primary_owner_required
        - participant_role_invalid
        - participant_exists
        - participant_not_found
        - reference_not_verified
        - artifact_not_owned
        - artifact_not_found
        - artifact_compensation_failed
        - material_type_not_allowed
        - material_too_large
        - material_empty
        - material_exists
        - material_not_found
        - version_conflict
        - invalid_item_transition
        - item_requirements_missing
        - item_content_required
        - item_blocked
        - item_terminal
        - assignee_not_authorized
        - dependency_exists
        - dependency_not_found
        - dependency_self
        - dependency_cycle
        - dependency_campaign_mismatch
        - dependency_terminal
        - artifact_exists
        - artifact_asset_mismatch
        - artifact_link_not_found
        - content_hash_mismatch
        - dependency_unavailable
        - dependency_invalid_response
        - rate_limited
        - internal_error
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, correlationId]
          properties:
            code: { $ref: '#/components/schemas/PublicErrorCode' }
            message: { type: string }
            correlationId: { type: string, minLength: 1, maxLength: 128 }
            details: {}
...